Implement a message content size adjustment in KryptonTaskDialog (#92)
* Fix placement of build artifacts of KryptonExplorer

* Adjust build location of Designer. Guard file execution in Explorer

* Adjust hint paths in all example projects (addition to issue #87)

* Implement a message content size adjustment

The content display control of the KryptonTaskDialog is changed
based on the size of the content.  If the content contains more
than 20 lines of text, the control changes to a multi-line text
container.  The size of the KryptonTaskDialog is also limitted
to 0.6 size of the display area to fix the issue of displaying
very large dialog boxes that don't fit in a screen.